Extract title header to TitleLegendHeader and use it in Funnel viz
